+++ app/controllers/admin/main_admin_controller.rb
@@ -6,6 +6,10 @@ module Admin
     layout 'assignment_content'
 
     def index
+      @dashboards = [{ name: t('resque.dashboard'), path: admin_resque_path }]
+      if Settings.rails_performance.enabled
+        @dashboards << { name: t('rails_performance.dashboard'), path: admin_performance_path }
+      end
       render :index
     end